Fix handling of skipped module, make it possible to force on command line.
When a module is marked as :skip in the build plan it was not possible
to force building it expect if a specific version or variant was passed
on the commande line:

   $ scbi --plan=xyz mod:v2

But it was not possible to force a skipped module to a default version.

   $ scbi --plan=xyz mod

This patch introduce a new specific version :force that can be used
to achieve that:

   $ scbi --plan=xyz mod:force

Add corresponding regression test and documentation.