Add support for barrier in group plan.
A barrier is a condition on the start of a group that will make
the group active or not.

   [ <op> <variant>
      ...
   ] /variant:version

Where <op> is = (equal) or /= (not equal) and variant a string that will
be checked into the build plan variant. The build plan variant is specified
with /<variant> after the build plan name.

   $ scbi --plan=dev/mpi ...

This will activate the following build plan:

   [ = mpi
      moduleA
      moduleB:v2
   ] /mpi

And so finally this is equivalent to:

   moduleA/mpi
   moduleB/mpi:v2

Add corresponding documentation and regression test.