Make OS dependencies part of the plug-in build id.
This ensure that if an external dependency is change the build
will be done. Then we can remove the hard coded dependency on
gcc. This is not correct anyway as some softwares can be built
with another compiler than GCC (or no compiler at all Python).

Add corresponding regression test.