Rework the mingw header and runtime support.
These were old plug-in and needed some updates to better set the
dependencies and avoiding copying some date from a plug-in to
another on which it depends. It is better to keep each plug-in
self contained when possible.

Also some OS dependencies have been added to the binutils and
partial gcc compiler needed for cross-compilation bootstrap.