Generate SCBI_MODULES_DIR associative array.
This contains the installation directory of all dependent modules.
It a a way to document the build environment and at the same time
it could be useful in complex cases where one module need to get
some specific data from a dependency.

Update regression test07 & test08 to check for this new variable.