Rework the dependencies handling.
A module gets the environment of:

1. The direct build-depends and recursively the depends
2. The self build-env,
3. The recursive environment of depends.

This is almost equivalent to previous version except for 1 where we
get first the environment of build-depends direct dependencies and then
the environment of the standard dependencies. That is, at the end we
have a smaller (but safer and cleaner) environment set.

Add corresponding regression test.