Do late resolution of the native auto-variant version.
Doing this during the creation of with-variant was not correct as it
was imposible the same version to all usages. It is perfectly correct
to have a specific version when used for a specific module.

Having for example in the build plan:

   module1:v1.15
   some-module=module1:v1.20

Means that v1.20 must be used when using as dependency of some-module
otherwise v1.15 can be used.

Add corresponding regression test.