Use parent module name for alone auto variant.
If only the auto variant is specified and we have a for-module we
use it as the name of the variable instead of defaulting to "default"
variant. This ensures that a specific version will be built into
a separate directory (as all variants). So, different versions for
different modules won't clash by using the same build directory.

We do that only if only auto variant is specified that let users
specified a specific build variant to use if needed and so sharing
the same build for different module is needed.

Add corresponding regression test.