Add support for only-explicit-build hook.
This hook is optional and when defined it must return true or false.
If true is returned then the module will be built only if specified
explicitly on the command line. Otherwise the build will be skip. Yet,
when found on dependencies it will be used to setup the env. This is
designed to be used with plug-in building large module which can be
slow and where building is not mandatory (like a plug-in to build
a compiler).