scbi: Allows for optional arguments for tests hook.
That is, it is possible to pass some option to the test hook, this
is often needed to specifiy one or mutiple tests to be checked.

For example, build and pass a single test:

  $ scbi --tests=0001_turl c-aws

or test only all tests having prefix 001:

  $ scbi --tests:only="001*" c-aws

Of course this option has a semantic specific to the project it
applies.

Add regression test and documentation.