scbi: Refine the meta-modules' modules handling.
The meta-modules' modules where iterated only if --deps was used. This
was done on purpose to be able to run just the setup/wrapup hooks.

We now iterate the modules if there is no setup/wrapup hooks. This makes
it possible to pass on the command line a simple meta-module (alias kind)
listing some modules to handle and yet process them.

That is:

   $ scbi meta:dev

is equivalent to:

   $ scbi module1:dev module2:dev module3:dev

if hook meta-modules contains the three listed modules.

Add corresponding regression test.