Rework auto-variant to get the version of the OS if native is used.
This is cleaner as the displayed version and version in the build plan
for native plug-in is the one found in the OS. So when native is selected
for an auto variant we know which version has actually been selected.

Some simplification have been done which should make this part
easier to understand and maintain. All status/error values have
been renamed and set into 5_tools to be shared.

Adjust regression tests accordingly.