Implement new reference "keep" meaning no-source update.
When a module is given the keep version on the command line the
sources are not updated (so no config, build, install nor setup, wrapup
steps are conducted). But the module is not ignored (as for "skip") and
will still be used as if up-to-date even if the sources have changed
since it was first compiled.

Add regression tests and corresponding documentation.