Add a cache for the plan for speed-up.
This is especially usefull on Windows (Mingw32) where the auto variant
resolution is slow due to the need to do lot of call to pacman.

A new option --clear-cache is introduced to reset/delete the cache and
force a full recomputation of the cache. This is needed as some cases
are not possible to cover (or too costly) as checking if a specific
version of a dependency has been hard coded (or changed) in a plug-in.

Add documentation and corresponding regression test.